+++ chainindexer.go
@@ -69,8 +69,8 @@ type addrIndexer struct {
 	addrIndexJobs   chan *indexBlockMsg
 	writeRequests   chan *writeIndexReq
 	progressLogger  *blockProgressLogger
-	currentIndexTip int64
-	chainTip        int64
+	currentIndexTip int32
+	chainTip        int32
 	sync.Mutex
 }
